Add functions to copy and insert Org IDs
The commit adds two functions to copy and insert Org IDs with
user-defined descriptions. The `my/copy-org-id-at-point` function
copies the ID property of the heading at point to the kill-ring.
The `my/insert-org-id-from-kill-ring` function inserts a link to an
Org ID from the kill-ring with a user-defined description. If the
description is empty, the function retrieves the heading from the
org-id database using `my-get-heading-from-org-id-db` function.