Add function to jump to org-mode ID links
This commit adds a new function `my/org-link-goto-at-point` which
checks if the link at point is a file link or an ID link, and jumps
to the appropriate location. The function also includes a helper
function `my-parse-link-id` to parse the ID from an org-mode link of
the form `id:xxxxxxxxxxxx`.