Update init.el and init-deps.el, add my-check-for-font function to init-pre.el
The changes in this commit update init.el and init-deps.el files and add
a new function called my-check-for-font to init-pre.el. The
my-check-fo
r-font function checks if a given font is available and prompts the
user
to continue if it's not found. This function is used in
init-display.el
to check for the Symbola font.