Refactor my/find-tags-file function and add optional arguments
This commit refactors the my/find-tags-file function in init-misc.el to
include two optional arguments: ask and tag-file-name. It also adds a
TO
DO comment for testing purposes. The function now recursively searches
e
ach parent directory for a file named TAGS or the specified
tag-file-nam
e. If the file is found, it returns the file path. If the buffer is
not
visiting a file or the file is not found, it returns nil.