Add my/kill-buffers-by-pattern function to utils.el
This commit adds the my/kill-buffers-by-pattern function to utils.el. Th
is function allows the user to kill buffers whose names match a
specifie
d pattern. The function prompts the user for a pattern, searches
through
all buffers, and kills the buffers whose names match the pattern. The
p
attern is a regular expression that is compared against buffer names
usi
ng 'string-match-p'.