Refactor init-display.el to hide list of minor modes in mode-line
This commit refactors the init-display.el file to hide the list of minor
modes in the mode-line. It adds a new function `my/purge-minor-modes`
that
removes the specified minor modes from the `minor-mode-alist`. The
func
tion is then added to the `after-change-major-mode-hook` to be
executed
after each major mode change. This improves the readability and
declutte
rs the mode-line.